Move redundant information and do a big rework of contribution guide
