+++ pyomo/devel/example/bar.py
