+++ pyomo/devel/example/plugins/__init__.py
@@ -9,17 +9,9 @@
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
 
-from pyomo.core.base import Transformation, TransformationFactory
+# Define a 'load()' function, which simply imports
+# sub-packages that define plugin classes.
 
 
-@TransformationFactory.register(
-    'contrib.example.xfrm',
-    doc="An example of a transformation in a pyomo.contrib package",
-)
-class Xfrm_PyomoTransformation(Transformation):
-    def __init__(self):
-        super(Xfrm_PyomoTransformation, self).__init__()
-
-    def create_using(self, instance, **kwds):
-        # This transformation doesn't do anything...
-        return instance
+def load():
+    from pyomo.devel.example.plugins import ex_plugin
