+++ pyomo/devel/example/tests/__init__.py
@@ -9,9 +9,4 @@
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
 
-# Define a 'load()' function, which simply imports
-# sub-packages that define plugin classes.
-
-
-def load():
-    from pyomo.contrib.example.plugins import ex_plugin
+# Tests for pyomo.devel.example
