+++ pyomo/environ/__init__.py
@@ -40,7 +40,6 @@ _packages = [
     'pyomo.contrib.community_detection',
     'pyomo.contrib.cp',
     'pyomo.contrib.cspline_external',
-    'pyomo.contrib.example',
     'pyomo.contrib.fme',
     'pyomo.contrib.gdp_bounds',
     'pyomo.contrib.gdpopt',
@@ -55,6 +54,7 @@ _packages = [
     'pyomo.contrib.simplification',
     'pyomo.contrib.solver',
     'pyomo.contrib.trustregion',
+    'pyomo.devel.example',
 ]
